CSS Revision to Doctor Who site
Prepped HTML for CSS, modified links so Doctor pages link back and jump
to exact doctor on homepage, removed the 1” border attribute from
original table, and added a CSS styles page with rules for body, tf, a,
p, a:hover, h1, and img.  I also created a descendant rule so that all
the borders of the 150px pictures on the index page are in uniform and
different from the borders of the 300px doctor images on their
individual pages.